Server/requirements.txt
flask
flask-cors
google-generativeai
python-dotenv
python-dateutil
pandas
scikit-learn
plotly

